Changes for PUT /forms/:id
Previously for sharing a new survey, we only accepted a user list. Now we
accept a user list and an email list. For the emails, we check which ones
correspond to existing users. For existing users, we give a notification;
for new emails, we send an email.

Removed some dead code too.