When polling for notifications every 30 seconds, only fetch new ones
When the page is first loaded, all notifications are fetched. Then when
the app polls for updates, it only polls for notifications newer than
what it already has.

This is for performance, to reduce DB queries. We need up to 2 DB queries
per notification, i.e. 100 queries every 30 seconds (if you have 50
notifications), even when the browser tab is just sitting idle. Now it is
only 1 query per 30 seconds.

The downside is, if a notification is read or deleted (i.e. you have two
tabs open), then your notification view will be stale until you reload the
page.