Do not do full page refresh after login
A full page fresh includes re-downloading and parsing bundle.js and takes
several seconds. While it is loading, the browser shows a broken page
(the home page, but all AJAX requests are aborted due to the refresh).

Now we just go right to the home page, and the navbar updates.