Change routes for groups. Was /?group=X, now is /class/X or /studentorg/X
This looks better. It makes sense to have the group ID in the path but to
continue to use query params for share tokens. This gets around a bug with
query params too.

Originally we used query params so that / and /?group=X would be the same
route. Then you could navigate between those without reloading all of the
main components (homepage, feedpage, sidebar, navbar). Now we will need to
do a reload, but we resolved the main issue -- the sidebar. Now the
sidebar keeps its data statically, so the user won't notice the sidebar
being empty when navigating between pages.